+++ NEWS.md
@@ -133,6 +133,9 @@ Red Panda C++ Version 2.27
   - fix: Can't find symbols indirectly included by other files.
   - enhancement: Function tip's width changes with editor width.
   - fix: '<' / '>' not shown in function tips.
+  - enhancement: In debug console, Ctrl+C/Ctrl+X/Ctrl+V conflicts with application action.
+  - enhancement: Auto hide Edit/Selection/Code/Refactor menu if no file openning.
+  - enhancement: Auto hide Project menu if no project openning.
     
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.